requirements.txt
google
google-genai
streamlit
pillow